pyproject.toml
[tool.poetry]
name = "pinecone-local-dev-starter"
version = "0.1.0"
description = ""
authors = ["magisystem0408 <masato.matsudo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.1"
pinecone-client = "^3.1.0"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
